+++ module/Information/src/Information/Entity/Db/InformationLangue.php
@@ -2,7 +2,9 @@
 
 namespace Information\Entity\Db;
 
-class  InformationLangue {
+class InformationLangue
+{
+    const ID_LANGUE_PAR_DEFAUT = 'FR';
 
     /** @var string */
     private $id;
@@ -13,25 +15,16 @@ class  InformationLangue {
     /** @var string */
     private $drapeau;
 
-    /**
-     * @return string
-     */
     public function getId(): string
     {
         return $this->id;
     }
 
-    /**
-     * @return string
-     */
     public function getLibelle(): string
     {
         return $this->libelle;
     }
 
-    /**
-     * @return string
-     */
     public function getDrapeau(): string
     {
         return $this->drapeau;