+++ module/Information/src/Information/Service/InformationLangue/InformationLangueServiceAwareTrait.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Information\Service\InformationLangue;
+
+trait InformationLangueServiceAwareTrait
+{
+    protected InformationLangueService $informationLangueService;
+
+    public function setInformationLangueService(InformationLangueService $informationLangueService): void
+    {
+        $this->informationLangueService = $informationLangueService;
+    }
+}